fixed tests for bank account details
